requirements.txt
asttokens==3.0.0
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.1.0
blinker==1.8.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
entrypoints==0.3
executing==2.2.0
fastjsonschema==2.21.1
flasgger==0.9.7.1
flask==3.0.3
Flask-JWT-Extended==4.6.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==8.5.0
incremental==16.10.1
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.2
more-itertools==4.2.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
netifaces==0.10.4
oauthlib==3.1.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt-toolkit==3.0.50
pure-eval==0.2.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygments==2.19.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
pyzmq==26.3.0
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
soupsieve==2.6
ssh-import-id==5.10
stack-data==0.6.3
systemd-python==234
tinycss2==1.2.1
tornado==6.4.2
traitlets==5.14.3
Twisted==18.9.0
typing-extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.0.6
yarg==0.1.9
zipp==3.20.2
zope.interface==4.7.1
